refactor: optimize code generation logic
