+++ packages/vue-generator/src/templates/vue-template/templateFiles/src/mcp/App.vue.mcp-onMounted.ts
@@ -1,0 +1,4 @@
+export default () => {
+  return `  createProxyTransport();
+  initMcpServer();`
+}
