+++ packages/vue-generator/src/templates/vue-template/templateFiles/src/mcp/App.vue.mcp-style.ts
@@ -1,0 +1,6 @@
+export default () => {
+  return `
+:deep(.next-sdk-trigger-btn) {
+  font-size: 30px;
+}`
+}
