+++ packages/vue-generator/src/templates/vue-template/templateFiles/src/mcp/App.vue.mcp-template.ts
@@ -1,0 +1,3 @@
+export default () => {
+  return `    <TinyRemoter :sessionId="SESSION_ID" class="remoter" />`
+}
