+++ packages/vue-generator/src/templates/vue-template/templateFiles/src/mcp/base.ts
@@ -1,0 +1,6 @@
+export default (schema, options) => {
+  const { agentRoot, sessionId } = options
+
+  return `export const AGENT_ROOT = '${agentRoot}'
+export const SESSION_ID = '${sessionId}'`
+}
